Character animation testing
Added Animation Starter Pack assets and remapped some of the assets for use in ue5. Created a blend space to blend idle animation to jog animation based on velocity, and added this to an animation blueprint (this replaced the original third person animation blueprint). All these items can be found in Content\Animations